+++ packages/addfox/CHANGELOG.md
@@ -1,5 +1,11 @@
 # addfox
 
+## 0.2.6
+
+### Patch Changes
+
+- @addfox/cli@0.2.6
+
 ## 0.2.5
 
 ### Patch Changes